api/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
Pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
numpy==1.24.2
ultralytics==8.0.20